Created `ServiceNameField`; implemented use of `FieldSubtype`; moved field subclasses to a separate file (`base.py` getting too long)
